pyproject.toml
[project]
name = "pdf2md"
version = "0.1.0"
description = "Web service that converts any publicly-accessible PDF to markdown via URL rewrite"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.6.0",
]

[project.optional-dependencies]
# marker-pdf pulls in torch (~2GB+); install separately in production
ml = [
    "marker-pdf>=1.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pdf2md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "respx>=0.22.0",
]
